+++ crafters/nlp/DeepSegmenter/README.md
@@ -1,0 +1,3 @@
+# DeepSegmenter
+
+Designed with ASR outputs in mind, DeepSegment uses BiLSTM + CRF for automatic sentence boundary detection. It significantly outperforms the standard libraries (spacy, nltk, corenlp ..) on imperfect text and performs similarly for perfectly punctuated text.
