+++ crafters/nlp/DeepSegmenter/config.yml
@@ -1,0 +1,3 @@
+!DeepSegmenter
+with:
+  {}
