+++ crafters/nlp/DeepSegmenter/requirements.txt
@@ -1,0 +1,1 @@
+deepsegment
