+++ crafters/nlp/DeepSegmenter/tests/test_deepsegmenter.py
@@ -1,0 +1,9 @@
+from .. import DeepSegmenter
+
+
+class TestClass:
+    def test_deepsegmenter(self):
+        """here is my test code
+
+        https://docs.pytest.org/en/stable/getting-started.html#create-your-first-test
+        """
