+++ crafters/nlp/SlidingWindowSegmenter/README.md
@@ -1,0 +1,3 @@
+# SlidingWindowSegmenter
+
+split the text on the doc-level into overlapping substrings on the chunk-level.
