+++ crafters/nlp/SlidingWindowSegmenter/config.yml
@@ -1,0 +1,3 @@
+!SlidingWindowSegmenter
+with:
+  {}
