+++ crafters/nlp/SlidingWindowSegmenter/manifest.yml
@@ -1,0 +1,11 @@
+manifest_version: 1
+name: SlidingWindowSegmenter
+type: crafter
+description: split the text on the doc-level into overlapping substrings on the chunk-level.
+author: Jina AI Dev-Team (dev-team@jina.ai)
+url: https://jina.ai
+vendor: Jina AI Limited
+documentation: https://github.com/jina-ai/jina-hub
+version: 0.0.1
+license: apache-2.0
+keywords: [Some keywords to describe the executor, separated by commas]
