+++ crafters/nlp/SlidingWindowSegmenter/requirements.txt
@@ -1,0 +1,2 @@
+Python packages required to install
+ separated by commas
